Add funroll-loops flags to just the binary
This patch make the spec to build twice, once without the funroll-loops
optimizations and the other with them. In the end the binaries are the ones
that are overwrite. This is to keep the performance in the binaries at the
same time that the libraries aloud the compilation of other packages ( with the
perf optimization flags some libraries do not have all the functionality other
package request)

Signed-off-by: vrodri3 <victor.rodriguez.bahena@intel.com>