tcl: Remove extra install command from install_append
extra `%make_install install-binaries` in install_append conflicts with
autospec's current approach to handle excludes, as rm is executed before
'install-binaries', which results in rpm build error due to sqlite3_analyzer
installed but not packaged. This is a temporary workaround since the
extra make install was done to install optimized binaries.